Backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
